add: 画面スリープ削除時に、ウィジェットを更新させる処理追加
